AC tests, clean up AC reference annual energy use code
