--- openfisca_nsw_safeguard/tests/ESS/ESS_D16/ESS_D16_cooling_and_heating_hours.yaml
@@ -1,16 +1,0 @@
-- name: test ESS AC cooling annual energy use
-  period: 2021
-  absolute_error_margin: 0.1
-  input:
-      ESS_D16__cooling_power_input: [1, 1, 1]
-      Appliance__zone_type: [hot, average, cold]
-  output:
-      ESS_D16__cooling_annual_energy_use: [1274, 429, 285]
-- name: test ESS AC heating annual energy use
-  period: 2021
-  absolute_error_margin: 0.1
-  input:
-      ESS_D16__heating_power_input: [1, 1, 1]
-      Appliance__zone_type: [hot, average, cold]
-  output:
-      ESS_D16__heating_annual_energy_use: [109, 648, 1534]
