+++ openfisca_nsw_safeguard/tests/ESS/ESS_HEAB/ESS_HEAB_AC/ESS_HEAB_install_new_AC_eligibility_requirements.yaml
@@ -1,0 +1,157 @@
+- name: OpenFisca Test forimplementation_requirementsfor activity F4
+  period: 2022
+  absolute_error_margin: 0
+  input:
+    PDRS_activity_type: [
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              install_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+              replace_AC,
+          ]
+    BCA_building_class: [
+              BCA_Class_1a,
+              BCA_Class_1b,
+              BCA_Class_2,
+              BCA_Class_3,
+              BCA_Class_4,
+              BCA_Class_5,
+              BCA_Class_6_shop,
+              BCA_Class_6_cafe_or_rest,
+              BCA_Class_7a,
+              BCA_Class_7b,
+              BCA_Class_8,
+              BCA_Class_9a_clinic,
+              BCA_Class_9a_hospital,
+              BCA_Class_9b_schools,
+              BCA_Class_9b_theatres,
+              BCA_Class_10a,
+              BCA_Class_10b,
+              BCA_Class_10c,
+              BCA_Class_1a,
+              BCA_Class_1b,
+              BCA_Class_2,
+              BCA_Class_3,
+              BCA_Class_4,
+              BCA_Class_5,
+              BCA_Class_6_shop,
+              BCA_Class_6_cafe_or_rest,
+              BCA_Class_7a,
+              BCA_Class_7b,
+              BCA_Class_8,
+              BCA_Class_9a_clinic,
+              BCA_Class_9a_hospital,
+              BCA_Class_9b_schools,
+              BCA_Class_9b_theatres,
+              BCA_Class_10a,
+              BCA_Class_10b,
+              BCA_Class_10c,
+          ]
+  output:
+    ESS_is_not_residential_building: [
+              False,
+              False,
+              True,
+              True,
+              False,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              False,
+              False,
+              True,
+              True,
+              False,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+          ]
+    ESS_HEAB_install_or_replace_AC_meets_eligibility_requirements: [
+              False,
+              False,
+              False,
+              True,
+              False,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              False,
+              False,
+              False,
+              True,
+              False,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+              True,
+          ]
