+++ openfisca_nsw_safeguard/tests/ESS/ESS_HEAB/ESS_HEAB_AC/ESS_HEAB_install_new_AC_implementation_requirements.yaml
@@ -1,0 +1,20 @@
+- name: OpenFisca Test forimplementation_requirementsfor activity F4
+  period: 2022
+  absolute_error_margin: 0
+  input:
+    Equipment_is_installed: [
+      True,
+      False,
+      True
+    ]
+    implementation_is_performed_by_qualified_person: [
+      True,
+      True,
+      False
+    ]
+  output:
+    ESS_HEAB_install_or_replace_AC_meets_implementation_requirements: [
+      True,
+      False,
+      False
+    ]
