+++ openfisca_nsw_safeguard/variables/High_Efficiency_Appliances_Business (HEAB)/AirConditioner/ESS_HEAB_install_new_AC/ESS_HEAB_install_or_replace_AC_meets_implementation_requirements.py
@@ -1,0 +1,28 @@
+import numpy as np
+from openfisca_core.variables import Variable
+from openfisca_core.periods import ETERNITY
+from openfisca_core.indexed_enums import Enum
+from openfisca_nsw_base.entities import Building
+
+
+class ESS_HEAB_install_or_replace_AC_meets_implementation_requirements(Variable):
+    value_type = bool
+    entity = Building
+    default_value = False
+    definition_period = ETERNITY
+    label = 'Does the implementation meet all of the Implementation' \
+            ' Requirements defined in installing a high efficiency air conditioner for Business?'
+    metadata = {
+        'alias': "HEAB AC Install meets implementation requirements",
+    }
+
+    def formula(buildings, period, parameters):
+        is_installed = buildings(
+            'Equipment_is_installed', period)
+        performed_by_qualified_person = buildings(
+            'implementation_is_performed_by_qualified_person', period)
+
+        return (
+                is_installed * 
+                performed_by_qualified_person
+                )
