--- openfisca_nsw_safeguard/variables/High_Efficiency_Appliances_Business (HEAB)/AirConditioner/ESS_HEAB_install_new_AC/F4_eligibility_requirements.py
@@ -1,4 +1,0 @@
-from openfisca_core.variables import Variable
-from openfisca_core.periods import ETERNITY
-from openfisca_core.indexed_enums import Enum
-from openfisca_nsw_base.entities import Building
